Initial gitleaks configuration file
This configuration file is built from combining
https://github.com/upenn-libraries/omeka_classic/blob/master/.gitleaks.toml
and
https://github.com/NOAA-EDAB/NEesp/blob/main/.github/workflows/.gitleaks.toml
which were found on GitHub. These cover the `leaky-repo.toml` example from
gitleaks as well as expanding on the rules.

Some of these rules were either removed, commented out, or adjusts to reduce the
level of false positives after doing some test runs. We may want to revisit this
in the future if we feel like some things are being missed.